First approach rethinking test structure
